pyproject.toml
[project]
name = "physical-mcp"
version = "0.1.0"
description = "Ambient perception MCP server — connect cameras to LLMs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.0",
    "opencv-python>=4.8",
    "numpy>=1.24",
    "Pillow>=10.0",
    "imagehash>=4.3",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "click>=8.0",
    "qrcode>=7.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.30"]
google = ["google-genai>=1.0"]
all = ["anthropic>=0.40", "openai>=1.30", "google-genai>=1.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "ruff>=0.1"]

[project.scripts]
physical-mcp = "physical_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/physical_mcp"]

[tool.ruff]
target-version = "py310"
